Show changes issues evidence datatable

+++ app/assets/javascripts/tylium/modules/issues/evidence.js.coffee
@@ -1,5 +1,5 @@
 document.addEventListener "turbolinks:load", ->
-  if $('body.issues.show').length
+  if $('body.evidence.new').length
     $('#evidence-host-list a[data-toggle~=pill]').on 'click', (ev)->
       path   = $(this).data('path')
       node   = $(this).data('node')
@@ -11,16 +11,21 @@ document.addEventListener "turbolinks:load", ->
           .then (html) ->
             $("##{node}").html(html)
 
-    $('[data-behavior~=add-evidence-container]').on 'change', '#evidence_content', ->
-      $('#template-content').text($(this).val())
+    $nodeListOptions = $('[data-behavior~=existing-node-list] option')
 
     $('[data-behavior~=add-evidence-container]').on 'keyup', '#evidence_node', ->
       rule = new RegExp($(this).val(), 'i')
-      $('[data-behavior~=existing-node-wrapper]').hide();
-      $('[data-behavior~=existing-node-wrapper]').filter ->
-        rule.test($(this).find($('[data-behavior~=existing-node-label]')).text())
+      $nodeListOptions.hide();
+      $nodeListOptions.filter ->
+        rule.test($(this).text())
       .show()
 
+    $nodeListOptions.mousedown (e) ->
+      e.preventDefault()
+      $(this).prop 'selected', !$(this).prop('selected')
+      $(this).parent().focus()
+
+  if $('body.issues.show').length
     $table = $('[data-behavior~=dradis-datatable]')
     $table.on 'dradis:datatable:bulkDelete', ->
       $('[data-behavior~=evidence-count]').text($table.DataTable().rows().count());
